+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ### Version 0.0.5
+- Ability to set Bundle Id, Version Number, Build Number and Display Name of app
 - Handle when a user uses spaces in an app name when creating a project
 
 ### Version 0.0.4